Added test for -n|--dry-run and implemented it.
Now the -n or --dry-run option will output a 'theoretical' renaming of
files, but will not make any changes to the filesystem. This is useful
when, for example, one is checking if the files are acceptable by the
script.

Added output to signify the dry run. Also refactored test to clean up
after itself, and refactored non-DRY test.